refactor: move the JSON-to-Protobuf converters out of the foundational package
Abblix.Utils is what every other package here builds on, and its own description offers it
as usable on its own - yet it exposed a converter whose signatures name Protocol Buffers
types. Anyone referencing it for the string, URI or caching helpers restored
Google.Protobuf and took the storage serialisation format into their contract, so
replacing that format would have broken consumers who never touched it.

The converters move beside the mappers that are their only callers and become internal
there, free to change with the format they serve. Google.Protobuf leaves Abblix.Utils
altogether, which is the larger half: the dependency was reaching every consumer of the
foundational package, storage or no storage.

This removes a public type from a package with fifteen published versions, and the risk was
weighed rather than assumed: the type is a plain conversion between two libraries' JSON
shapes, with no dependency on anything here, so a caller copies it. The generated message
types have the same problem and are not touched, because 25 of them are public in the
published assembly and that is a scale where a quiet removal is the wrong default - issue
#308, milestone 3.0.

The test moves with the code, into the assembly that can now see it, and its remark about
"retained public API" is corrected: what those overloads are is now our own internal
question, not a compatibility obligation.
