+++ src/Abblix.Oidc.Server/Features/Storages/Proto/Mappers/RequestedClaimsMapper.cs
@@ -20,7 +20,6 @@
 // CONTACT: For license inquiries or permissions, contact Abblix LLP at
 // info@abblix.com
 
-using Abblix.Utils.Json;
 
 namespace Abblix.Oidc.Server.Features.Storages.Proto.Mappers;
